+++ programs/registry/src/instructions/reopen_account.rs
@@ -51,7 +51,7 @@ pub fn reopen_account(accounts: &[AccountInfo], instruction_data: &[u8]) -> Prog
     create_account_with_signer(
         sender,
         user_account,
-        UserAccount::get_space(max_data_size) as u64,
+        UserAccount::get_space(max_data_size),
         &seeds!(SEED::USER_ACCOUNT, user_seed_id, &[user_id.account_bump]),
         &crate::ID,
     )?;
